requirements.txt
appdirs==1.4.4
bidict==0.21.3
certifi==2021.10.8
charset-normalizer==2.0.7
filelock==3.3.1
idna==3.3
python-engineio==4.2.1
python-socketio==5.4.1
requests==2.26.0
uritools==3.0.2
urlextract==1.4.0
urllib3==1.26.7
websocket-client==1.2.1
